오류수정
오류수정

+++ src/main/java/kr/co/kh/model/vo/ChartVO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 public class ChartVO {
     private String label;    // 요일, 월 등 (ex. 월, 화, 1월, 2월 등)
-    private Integer value;   // 카운트 값 (매출, 방문자 수 등)
+    private Double value;   // 카운트 값 (매출, 방문자 수 등)
 }
